# src/config/swagger.yaml
openapi: 3.0.0
info:
  title: API Documentation
  description: Descripción de la API
  contact:
    email: tidy@tidy.com
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo
  - url: http://localhost:5173
    description: Servidor del frontend
  - url: http://tidy-app.me
    description: Ruta de dominio para enviar correos electrónicos
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Subject:
      type: object
      properties:
        id:
          type: integer
          description: ID de la materia
        name:
          type: string
          description: Nombre de la materia
        description:
          type: string
          description: Descripción de la materia
        name_teacher:
          type: string
          description: Nombre del profesor
        userId:
          type: integer
          description: ID del usuario
...
security:
  - bearerAuth: []
paths:
...
  /subjects:
    get:
      summary: Obtener todas las materias del usuario autenticado
      description: Permite obtener todas las materias asociadas al usuario autenticado.
      operationId: getSubjects
      tags:
        - Subjects
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de materias obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subject'
        '401':
          description: No autorizado

    post:
      summary: Crear una nueva materia para el usuario autenticado
      description: Permite crear una nueva materia asociada al usuario autenticado.
      operationId: createSubject
      tags:
        - Subjects
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subjectName:
                  type: string
                  description: Nombre de la materia
                description:
                  type: string
                  description: Descripción de la materia
                name_teacher:
                  type: string
                  description: Nombre del profesor
      responses:
        '201':
          description: Materia creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subject'
        '400':
          description: Error en la solicitud
        '401':
          description: No autorizado
  /subjects/{id}:
    put:
      summary: Editar una materia del usuario autenticado
      description: Permite editar una materia asociada al usuario autenticado.
      operationId: updateSubject
      tags:
        - Subjects
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subjectName:
                  type: string
                  description: Nombre de la materia
                description:
                  type: string
                  description: Descripción de la materia
                name_teacher:
                  type: string
                  description: Nombre del profesor
      responses:
        '200':
          description: Materia actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subject'
        '400':
          description: Error en la solicitud
        '401':
          description: No autorizado

    delete:
      summary: Eliminar una materia del usuario autenticado
      description: Permite eliminar una materia asociada al usuario autenticado.
      operationId: deleteSubject
      tags:
        - Subjects
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Materia eliminada exitosamente
        '400':
          description: Error en la solicitud
        '401':
          description: No autorizado
...